Fixes issue #2431 - Add Jacoco check to logging extension
