+++ extension/standard/src/main/java/cloud/piranha/extension/standard/StandardExtension.java
@@ -31,6 +31,7 @@ import cloud.piranha.core.api.WebApplicationExtension;
 import cloud.piranha.core.api.WebApplicationExtensionContext;
 import cloud.piranha.extension.apache.fileupload.ApacheMultiPartExtension;
 import cloud.piranha.extension.herring.HerringExtension;
+import cloud.piranha.extension.logging.LoggingExtension;
 import cloud.piranha.extension.mimetype.MimeTypeExtension;
 import cloud.piranha.extension.scinitializer.ServletContainerInitializerExtension;
 import cloud.piranha.extension.security.servlet.ServletSecurityExtension;
@@ -39,7 +40,6 @@ import cloud.piranha.extension.servletannotations.ServletAnnotationsExtension;
 import cloud.piranha.extension.standard.annotationscan.StandardAnnotationScanExtension;
 import cloud.piranha.extension.standard.async.StandardAsyncExtension;
 import cloud.piranha.extension.standard.localeencoding.StandardLocaleEncodingExtension;
-import cloud.piranha.extension.standard.logging.StandardLoggingExtension;
 import cloud.piranha.extension.standard.policy.StandardPolicyExtension;
 import cloud.piranha.extension.tempdir.TempDirExtension;
 import cloud.piranha.extension.wasp.WaspExtension;
@@ -58,7 +58,7 @@ public class StandardExtension implements WebApplicationExtension {
     public void extend(WebApplicationExtensionContext context) {
         context.add(StandardAsyncExtension.class);
         context.add(StandardLocaleEncodingExtension.class);
-        context.add(StandardLoggingExtension.class);
+        context.add(LoggingExtension.class);
         context.add(MimeTypeExtension.class);
         context.add(StandardPolicyExtension.class);
         context.add(TempDirExtension.class);
